Create L_lang Grammer.txt
Part of our grammar, not finished yet, still need to add like ARITHMETIC OPERATORS and RELATIONAL OPERATOR. Might need to adjust this to fit with the correct token. One thing is we might need to add a bool variable in, because how are we gonna identify the conditions in while, do loops? If someone could check would be great.